3DiceCeeLo
This is my first version of the popular East Coast dice throwing game,
Cee Lo. Upon loading, the game prompts the user(s) for the number of
players (n) and the amount that each player will bet. The game then
simulates n dice throws and in the end, the winner(s) will be declared.